Refactor save() to take a Writer argument
The struct NpyOut is now used for appendable .npy files only and new
functions writeHeader() and writeData() live at module-level. Data with
known length can now be written to a non-seekable strean which might be
used in NpzOut later. Generating the header twice to estimate its length
is no longer necessary in allocateSave() and NpyOut.